Php Install Process Documentation
The resources folder is originally not included in the
`createDefault.ps1` file which is my bad. I should have created a
folder to hold the images.

the `topic.md` and `example.md` is kind of irrelevant or redundant in
some cases(topics about php), so I will likely remove these folders in
the future.

The `resources.md` file will most-likely have redundant content- so
will most likely remove later as well.