Unstable: Display Purchase results
I managed to present the results from the purchase process, but I
kind of messed up everything else.

The `get_html_elem(...param)` function causes inconsistent results
when displaying the content of the purchase process.

The `$_GET["param_key"]` causes errors when I try to retrieve the
transaction credentials for other items.

Though it works stably when there is only one `Item` for the
transaction process.