Major Changes: User Input Forms & code
The forms for receiving the user input is kind of still not working.
I don't really understand what the error message means.

I changed the code structure of the `index.php` file to support only
a single form for retrieving all the data from the user input.

I think this will help clear the contents of the code and the
view of the user. I want to scale down the code, as there are too many
moving parts and I don't know which of which is causing what problem.

I transferred some methods away from the `index.php` to the `store.php`
and removed some redundant comments inside the `item.php`.

I think I should review how an https request works

fuckccccccccccccccc stillll won'ttt worrrkr